Refactor App.tsx and upload route for clearer separation of concerns
Split the 420-line god component App.tsx into three focused modules:
- App.tsx (38 lines): auth boundary and session routing only
- AuthenticatedApp.tsx: hook orchestration with clear section grouping
- useAccountActions.ts: account management handlers as a proper hook

Extract per-file upload processing into a dedicated UploadService:
- uploadService.ts: metadata resolution, directory setup, deduplication
- uploadRoutes.ts: slim orchestration, validation, and response assembly

All 79 unit tests and 3 E2E tests pass with no functional changes.
